Prevent error in output when using JSON parameter
Previously, if `file` was not specified (and `json` was), an error would
be emitted about the file not being found. This was erroneous as only
the JSON input should have been considered. This applies a small fix to
remediate that, given by Wendell in usnistgov/OSCAL#1849.

I am unsure whether this is the desired long-term fix but I figured
opening a PR to memorialize the patch may be useful.

Co-authored-by: Wendell Piez <wendell.piez@nist.gov>