Integrate feedback from @david-waltermire:
1. Make second model with allow-other scaffold and include previous
model version with arch change, don't make the model changes logically
separate.
1. Add detail to conclusion paragraph and explicitly defined learning
objectives and that they were met.
1. Not explicit to review but caught when touching up models: make sure
the model version numbers in `<schema-version/>` in the tutorials bump
up with each change to make sense and signify a change. (NOTE: The bumps
are in patch number, not minor number, which is odd open review issue
to check that later).