Add allowed-value constraints tutorial (#508)
* Add allowed-value constraints tutorial for #292

* Intro and required values section for #292

Add introduction and section about how to defined mandatory, required
values for `allowed-values allow-other="no"` for the first segment of
the tutorial.

* Recommended values section and conclusion for #292

Add introduction and section about how to defined mandatory, recommended
values for `allowed-values allow-other="yes"` for the first segment of
the tutorial.

Added a conclusion for wrap-up given the current draft contents.

* Integrate feedback from @david-waltermire:

1. Make second model with allow-other scaffold and include previous
model version with arch change, don't make the model changes logically
separate.
1. Add detail to conclusion paragraph and explicitly defined learning
objectives and that they were met.
1. Not explicit to review but caught when touching up models: make sure
the model version numbers in `<schema-version/>` in the tutorials bump
up with each change to make sense and signify a change. (NOTE: The bumps
are in patch number, not minor number, which is odd open review issue
to check that later).